RD-16944 Moved week header code between components
