+++ src/components/CalendarMonth.jsx
@@ -9,6 +9,7 @@ import moment from 'moment';
 
 import { CalendarDayPhrases } from '../defaultPhrases';
 import getPhrasePropTypes from '../utils/getPhrasePropTypes';
+import noflip from '../utils/noflip';
 
 import CalendarWeek from './CalendarWeek';
 import CalendarDay from './CalendarDay';
@@ -25,6 +26,7 @@ import DayOfWeekShape from '../shapes/DayOfWeekShape';
 
 import {
   HORIZONTAL_ORIENTATION,
+  VERTICAL_ORIENTATION,
   VERTICAL_SCROLLABLE,
   DAY_SIZE,
 } from '../constants';
@@ -35,6 +37,7 @@ const propTypes = forbidExtraProps({
   horizontalMonthPadding: nonNegativeInteger,
   isVisible: PropTypes.bool,
   enableOutsideDays: PropTypes.bool,
+  numberOfMonths: PropTypes.number,
   modifiers: PropTypes.objectOf(ModifiersShape),
   orientation: ScrollableOrientationShape,
   daySize: nonNegativeInteger,
@@ -50,6 +53,7 @@ const propTypes = forbidExtraProps({
   firstDayOfWeek: DayOfWeekShape,
   setMonthTitleHeight: PropTypes.func,
   verticalBorderSpacing: nonNegativeInteger,
+  weekDayFormat: PropTypes.string,
 
   focusedDate: momentPropTypes.momentObj, // indicates focusable day
   isFocused: PropTypes.bool, // indicates whether or not to move focus to focusable day
@@ -65,6 +69,7 @@ const defaultProps = {
   horizontalMonthPadding: 13,
   isVisible: true,
   enableOutsideDays: false,
+  numberOfMonths: 1,
   modifiers: {},
   orientation: HORIZONTAL_ORIENTATION,
   daySize: DAY_SIZE,
@@ -85,6 +90,7 @@ const defaultProps = {
 
   // i18n
   monthFormat: 'MMMM YYYY', // english locale
+  weekDayFormat: 'dd',
   phrases: CalendarDayPhrases,
   dayAriaLabelFormat: undefined,
   verticalBorderSpacing: undefined,
@@ -150,6 +156,66 @@ class CalendarMonth extends React.PureComponent {
     this.captionRef = ref;
   }
 
+  isHorizontal() {
+    const { orientation } = this.props;
+    return orientation === HORIZONTAL_ORIENTATION;
+  }
+
+  isVertical() {
+    const { orientation } = this.props;
+    return orientation === VERTICAL_ORIENTATION || orientation === VERTICAL_SCROLLABLE;
+  }
+
+  renderWeekHeader(index) {
+    const {
+      daySize,
+      orientation,
+      weekDayFormat,
+      styles,
+      firstDayOfWeek,
+    } = this.props;
+    const { calendarMonthWidth } = this.state;
+    const verticalScrollable = orientation === VERTICAL_SCROLLABLE;
+    const horizontalStyle = {
+      left: index * calendarMonthWidth,
+    };
+    const verticalStyle = {
+      marginLeft: -calendarMonthWidth / 2,
+    };
+
+    let weekHeaderStyle = {}; // no styles applied to the vertical-scrollable orientation
+    if (orientation === HORIZONTAL_ORIENTATION) {
+      weekHeaderStyle = horizontalStyle;
+    } else if (this.isVertical() && !verticalScrollable) {
+      weekHeaderStyle = verticalStyle;
+    }
+
+    const header = [];
+    for (let i = 0; i < 7; i += 1) {
+      header.push((
+        <li key={i} {...css(styles.CalendarMonth_weekHeader_li, { width: daySize })}>
+          <small>{moment().day((i + firstDayOfWeek) % 7).format(weekDayFormat)}</small>
+        </li>
+      ));
+    }
+
+    return (
+      <div
+        {...css(
+          styles.CalendarMonth_weekHeader,
+          this.isVertical() && styles.CalendarMonth_weekHeader__vertical,
+          verticalScrollable && styles.CalendarMonth_weekHeader__verticalScrollable,
+          weekHeaderStyle,
+        )}
+        key={`week-${index}`}
+      >
+        <ul {...css(styles.CalendarMonth_weekHeader_ul)}>
+          {header}
+        </ul>
+      </div>
+    );
+  }
+
   render() {
     const {
       dayAriaLabelFormat,
@@ -174,12 +240,20 @@ class CalendarMonth extends React.PureComponent {
       renderMonthText,
       styles,
       verticalBorderSpacing,
+      numberOfMonths,
     } = this.props;
 
     const { weeks } = this.state;
     const monthTitle = renderMonthText ? renderMonthText(month) : month.format(monthFormat);
 
     const verticalScrollable = orientation === VERTICAL_SCROLLABLE;
+    const isHorizontal = this.isHorizontal();
+
+    const numOfWeekHeaders = this.isVertical() ? 1 : numberOfMonths;
+    const weekHeaders = [];
+    for (let i = 0; i < numOfWeekHeaders; i += 1) {
+      weekHeaders.push(this.renderWeekHeader(i));
+    }
 
     return (
       <div
@@ -208,6 +282,15 @@ class CalendarMonth extends React.PureComponent {
               {monthTitle}
             </strong>
           )}
+        </div>
+
+        <div
+          {...css(styles.CalendarMonth_weekHeaders,
+            isHorizontal && styles.CalendarMonth_weekHeaders__horizontal)}
+          aria-hidden="true"
+          role="presentation"
+        >
+          {weekHeaders}
         </div>
 
         <table
@@ -248,12 +331,48 @@ class CalendarMonth extends React.PureComponent {
 CalendarMonth.propTypes = propTypes;
 CalendarMonth.defaultProps = defaultProps;
 
-export default withStyles(({ reactDates: { color, font, spacing } }) => ({
+export default withStyles(({ reactDates: { color, font, spacing, zIndex } }) => ({
   CalendarMonth: {
     background: color.background,
     textAlign: 'center',
     verticalAlign: 'top',
     userSelect: 'none',
+  },
+
+  CalendarMonth_weekHeaders: {
+    position: 'relative',
+  },
+
+  CalendarMonth_weekHeader: {
+    color: color.placeholderText,
+    position: 'absolute',
+    top: -24,
+    zIndex: zIndex + 2,
+    textAlign: noflip('left'),
+  },
+
+  CalendarMonth_weekHeader__verticalScrollable: {
+    top: 0,
+    display: 'table-row',
+    borderBottom: `1px solid ${color.core.border}`,
+    background: color.background,
+    marginLeft: noflip(0),
+    left: noflip(0),
+    width: '100%',
+    textAlign: 'center',
+  },
+
+  CalendarMonth_weekHeader_ul: {
+    listStyle: 'none',
+    margin: '1px 0',
+    paddingLeft: noflip(0),
+    paddingRight: noflip(0),
+    fontSize: font.size,
+  },
+
+  CalendarMonth_weekHeader_li: {
+    display: 'inline-block',
+    textAlign: 'center',
   },
 
   CalendarMonth_table: {
