+++ src/components/CalendarMonthGrid.jsx
@@ -60,6 +60,7 @@ const propTypes = forbidExtraProps({
   isRTL: PropTypes.bool,
   transitionDuration: nonNegativeInteger,
   verticalBorderSpacing: nonNegativeInteger,
+  weekDayFormat: PropTypes.string,
 
   // i18n
   monthFormat: PropTypes.string,
@@ -98,6 +99,7 @@ const defaultProps = {
 
   // i18n
   monthFormat: 'MMMM YYYY', // english locale
+  weekDayFormat: 'dd',
   phrases: CalendarDayPhrases,
   dayAriaLabelFormat: undefined,
 };
@@ -265,6 +267,7 @@ class CalendarMonthGrid extends React.PureComponent {
       transitionDuration,
       verticalBorderSpacing,
       setMonthTitleHeight,
+      weekDayFormat,
     } = this.props;
 
     const { months } = this.state;
@@ -355,6 +358,8 @@ class CalendarMonthGrid extends React.PureComponent {
                 dayAriaLabelFormat={dayAriaLabelFormat}
                 verticalBorderSpacing={verticalBorderSpacing}
                 horizontalMonthPadding={horizontalMonthPadding}
+                numberOfMonths={numberOfMonths}
+                weekDayFormat={weekDayFormat}
               />
             </div>
           );
